website/requirements.txt
django
wagtail
psycopg2-binary
gunicorn
wagtail-metadata
djangorestframework